% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{getResult}
\alias{getResult}
\title{Get Result}
\usage{
getResult(odenet)
}
\arguments{
\item{odenet}{[\code{ODEnetwork}]\cr
Object of class \code{\link{ODEnetwork}}.}
}
\value{
a matrix with columns time and states 1 and 2 of class \code{\link{deSolve}}.
}
\description{
Getting result from numerical solving algorithm of given \code{\link{ODEnetwork}}.
}
\examples{
masses <- 1
dampers <- as.matrix(0.1)
springs <- as.matrix(4)
odenet <- ODEnetwork(masses, dampers, springs)
getResult(odenet)
}
