% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simuNetwork.R
\name{simuNetwork}
\alias{simuNetwork}
\title{Simulation of the Differential Equations}
\usage{
simuNetwork(odenet, times, ...)
}
\arguments{
\item{odenet}{[\code{ODEnetwork}]\cr
List of class \code{\link{ODEnetwork}}.}

\item{times}{[\code{numeric}]\cr
time sequence for which output is wanted,
the first value of \code{times} is the initial time.}

\item{...}{Additional arguments.}
}
\value{
an extended list of class [\code{\link{ODEnetwork}}].
}
\description{
Simulates the given \code{\link{ODEnetwork}} over a time range.
}
\examples{
masses <- 4:6
dampers <- diag(1:3)
springs <- diag(7:9)
odenet <- ODEnetwork(masses, dampers, springs)
position <- rep(10, 3)
velocity <- rep(0, 3)
odenet <- setState(odenet, position, velocity)
}

